web_app/requirements.txt
alabaster
alembic==1.7.4
aniso8601==9.0.1
appdirs==1.4.4
argh==0.26.2
argon2-cffi
arrow==0.13.1
astroid
async-generator
atomicwrites==1.4.0
attrs
autopep8
Babel
backcall
bcrypt
binaryornot
black==19.10b0
bleach
brotlipy==0.7.0
certifi==2021.5.30
cffi
chardet
charset-normalizer
click
cloudpickle
colorama
cookiecutter
cryptography
debugpy
decorator
defusedxml
diff-match-patch
dnspython==2.1.0
docutils
email-validator==1.1.3
entrypoints==0.3
flake8
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Dance==0.14.0
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
future==0.18.2
greenlet==1.1.2
idna
imagesize
importlib-metadata
importlib-resources==5.2.2
inflection==0.5.1
intervaltree
ipykernel
ipython
ipython-genutils
isort
itsdangerous==2.0.1
jedi
Jinja2==2.10
jinja2-time
jsonschema
jupyter-client
jupyter-core
jupyterlab-pygments
keyring
lazy==1.4
lazy-object-proxy
Mako==1.1.5
MarkupSafe==1.1.1
matplotlib-inline 
mccabe==0.6.1
mistune==0.8.4
mypy-extensions==0.4.3
nb-conda==2.2.1
nb-conda-kernels
nbclient
nbconvert
nbformat
nest-asyncio
notebook
numpydoc
oauthlib==3.1.1
packaging
pandocfilters
paramiko
parso==0.7.0
pathspec==0.7.0
pexpect
pickleshare
Pillow==8.4.0
pluggy
poyo
prometheus-client
prompt-toolkit
psutil
ptyprocess
pycodestyle
pycparser
pydocstyle
pyflakes
Pygments
PyJWT==1.4.2
pylint
pyls-black
pyls-spyder
PyNaCl
pyOpenSSL
pyparsing
pyrsistent
PySocks
python-dateutil
python-jsonrpc-server
python-language-server
python-slugify
pytz==2021.3
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.4.1
pyzmq
QDarkStyle
qstylizer
QtAwesome
qtconsole
QtPy
regex
requests
requests-oauthlib==1.3.0
rope
Rtree
Send2Trash
sip==4.19.13
six
snowballstemmer
sortedcontainers
Sphinx
sphinxcontrib-applehelp
sphinxcontrib-devhelp
sphinxcontrib-htmlhelp
sphinxcontrib-jsmath
sphinxcontrib-qthelp
sphinxcontrib-serializinghtml
spyder
spyder-kernels
SQLAlchemy==1.4.25
terminado==0.9.4
testpath
text-unidecode
textdistance
three-merge
tinycss
toml
tornado
traitlets
typed-ast
typing-extensions
ujson
Unidecode
urllib3==1.26.7
URLObject==2.4.3
watchdog
wcwidth
webencodings==0.5.1
Werkzeug==2.0.2
whichcraft
win-inet-pton
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
yapf
zipp
